fragment directory with classes added, working left drawer buttons with fragments changing fragments
