+++ app/src/main/java/com/greenfox/peridot/peridot_coz_android/activity/LoginActivity.java
@@ -3,10 +3,10 @@ package com.greenfox.peridot.peridot_coz_android.activity;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.support.annotation.NonNull;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.NavigationView;
 import android.support.design.widget.Snackbar;
+import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
@@ -21,6 +21,8 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.greenfox.peridot.peridot_coz_android.R;
+import com.greenfox.peridot.peridot_coz_android.fragment.FirstFragment;
+import com.greenfox.peridot.peridot_coz_android.fragment.SecondFragment;
 
 public class LoginActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
 
@@ -62,7 +64,6 @@ public class LoginActivity extends AppCompatActivity implements NavigationView.O
         });
 
 
-
         //for the nav bar on the left
 
         FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
@@ -84,7 +85,6 @@ public class LoginActivity extends AppCompatActivity implements NavigationView.O
         navigationView.setNavigationItemSelectedListener(this);
 
 
-
     }
 
     @Override
@@ -97,6 +97,7 @@ public class LoginActivity extends AppCompatActivity implements NavigationView.O
         return super.onCreateOptionsMenu(menu);
 
     }
+
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         int id = item.getItemId();
@@ -120,31 +121,58 @@ public class LoginActivity extends AppCompatActivity implements NavigationView.O
         return super.onOptionsItemSelected(item);
     }
 
-    public void saveData(View view){
+    public void saveData(View view) {
         SharedPreferences loginData = getSharedPreferences("userInfo", Context.MODE_PRIVATE);
         SharedPreferences.Editor editor = loginData.edit();
         editor.putString("userName", loginUsername.getText().toString());
         editor.putString("password", loginPassword.getText().toString());
         editor.commit();
-        Toast.makeText(this,"Saved", Toast.LENGTH_LONG).show();
+        Toast.makeText(this, "Saved", Toast.LENGTH_LONG).show();
     }
 
-    public void getData(View view){
-        Toast.makeText(this,"Displaying info",Toast.LENGTH_LONG).show();
+    public void getData(View view) {
+        Toast.makeText(this, "Displaying info", Toast.LENGTH_LONG).show();
         SharedPreferences loginData = getSharedPreferences("userInfo", Context.MODE_PRIVATE);
         String name = loginData.getString("userName", "");
-        String pw = loginData.getString("password","");
+        String pw = loginData.getString("password", "");
         String msg = "Saved User Name: " + name + "\nSaved Password: " + pw;
         dataView.setText(msg);
     }
 
-
-
-
-
-
     @Override
-    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-        return false;
+    public void onBackPressed() {
+        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        if (drawer.isDrawerOpen(GravityCompat.START)) {
+            drawer.closeDrawer(GravityCompat.START);
+        } else {
+            super.onBackPressed();
+        }
     }
+
+    @SuppressWarnings("StatementWithEmptyBody")
+    @Override
+    public boolean onNavigationItemSelected(MenuItem item) {
+        // Handle navigation view item clicks here.
+        int id = item.getItemId();
+
+        android.app.FragmentManager fragmentManager = getFragmentManager();
+
+
+        if (id == R.id.nav_first_layout) {
+            fragmentManager.beginTransaction()
+                    .replace(R.id.content_frame
+                    , new FirstFragment())
+                    .commit();
+        } else if (id == R.id.nav_second_layout) {
+            fragmentManager.beginTransaction()
+                    .replace(R.id.content_frame
+                            , new SecondFragment())
+                    .commit();
+        }
+
+        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
+        drawer.closeDrawer(GravityCompat.START);
+        return true;
+    }
+
 }
