+++ app/src/main/res/menu/activity_left_navigation_drawer.xml
@@ -3,21 +3,13 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_camera"
+            android:id="@+id/nav_first_layout"
             android:icon="@drawable/ic_menu_camera"
-            android:title="Import" />
+            android:title="First Layout" />
         <item
-            android:id="@+id/nav_gallery"
+            android:id="@+id/nav_second_layout"
             android:icon="@drawable/ic_menu_gallery"
-            android:title="Gallery" />
-        <item
-            android:id="@+id/nav_slideshow"
-            android:icon="@drawable/ic_menu_slideshow"
-            android:title="Slideshow" />
-        <item
-            android:id="@+id/nav_manage"
-            android:icon="@drawable/ic_menu_manage"
-            android:title="Tools" />
+            android:title="Second Layout" />
     </group>
 
     <item android:title="Communicate">
